feat: changed composition standards
previously this library generated a type with all of it interited types into a single type with properties merged.  This did not propertly represent the actual typescript structure and OpenAPI and Swagger 2.0 both support a composition style structure using "allOf".  This commit changes the standard for generating types.  Since its a dramatic shift from the current paradigm, I consider this a major change.

This commit also includes suport for new "ignoreProperties", "ignoreTypes", and "includeTypes" to help with including/excluding types and properties when referencing types that cant be annotated

BREAKING CHANGE